requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
langchain>=0.0.184
google-cloud-aiplatform>=1.36.0
google-generativeai>=0.3.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
notion-client>=1.0.0
python-multipart>=0.0.5
pytest>=7.0.0
black>=22.0.0
isort>=5.10.0
ruff>=0.0.290